requirements.txt
Django==1.5.1
PyYAML==3.10
Unipath==0.2.1
Werkzeug==0.8.3
django-analytical==0.15.0
django-extensions==1.1.1
django-intercom==0.0.6
django-mailchimp-v1.3==1.3.0
python-intercom==0.2.8
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
django-markdown-deux
South==0.8.2
git+https://github.com/ZmG/wsk_tutorial.git#egg=WskTutorial_dev
django-storages==1.1.8
boto==2.13.3
psycopg2==2.6.0
